mpd: Teach to create multiple daemon instances
This is a breaking change for current MPD configurations, but migration
is simple. A config that was previously

    services.mpd = {
      enable = true;
      extraConfig = ''
        audio_output {
          type "pulse"
	  name "PulseAudio"
        }
      '';
    };

is now

    services.mpd = {
      enable = true;
      daemons.default = {
        extraConfig = ''
          audio_output {
            type "pulse"
            name "PulseAudio"
          }
        '';
      };
    };

and the service name or default directories will not change.

This is accomplished through a `default` daemon option and a name check
that, while not super idiomatic to Nixpkgs as far as I can tell, works
well here and allows for simpler configuration logic (as opposed to, for
example, a mostly duplication of the config at the root level that
slightly differs because it's not in the extra daemon set that would be
more confusing to use and clutter the docs).